% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aro.clo.points.R
\name{aro.clo.points}
\alias{aro.clo.points}
\title{aro.clo.points}
\usage{
aro.clo.points(target, reference)
}
\arguments{
\item{target}{kxm matrix or object of class "mesh3d"}

\item{reference}{numeric: a kxm matrix (coordinates)}
}
\value{
position numeric: a vector of the row indices

distances numeric: a vector of the coordinates distances
}
\description{
Find the closest matches between a reference (2D or 3D matrix) and a target matrix (2D/3D) or mesh returning row indices and distances
}
\examples{
#load an example: mesh, and L set
data(yoda_sur)
data(yoda_set)
sur<-yoda_sur
set<-yoda_set
ver_pos<-aro.clo.points(target=sur,reference=set) 
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
