% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arothron.r
\docType{package}
\name{Arothron-package}
\alias{Arothron-package}
\alias{Arothron}
\title{R functions for Geometric Morphometrics Analyses}
\description{
Tools for Geometric Morphometrics. The package includes tools of Virtual Anthropology to build virtual cavities as endocasts, to align two disarticulated models belonging to the same specimen, and functions to import and export format files into another format.
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
