% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2list.R
\name{array2list}
\alias{array2list}
\title{array2list}
\usage{
array2list(array)
}
\arguments{
\item{array}{a kx3xn array with landmark coordinates}
}
\value{
a list containing the landmark configurations stored as separated elements
}
\description{
converts an array in a list storing each element of the third dimension of the array (specimen) as element of the list
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Paolo Piras, Pasquale Raia
}
