% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bary.mesh.R
\name{bary.mesh}
\alias{bary.mesh}
\title{bary.mesh}
\usage{
bary.mesh(mesh)
}
\arguments{
\item{mesh}{matrix mesh vertex}
}
\value{
barycenter numeric: x,y,z coordinates of the barycenter of the mesh
}
\description{
This function calculates the barycenter of a matrix or a 3D mesh
}
\examples{
#load an example: mesh, and L set
data(SCP1.mesh)
sur<-SCP1.mesh
bary<-bary.mesh(mesh=sur)
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
