% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Can_lup.R
\docType{data}
\name{Can_lup}
\alias{Can_lup}
\title{example dataset}
\usage{
data(Can_lup)
}
\description{
mesh coming from a ct-scan of Canis lupus
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
