% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec.curve.r
\name{dec.curve}
\alias{dec.curve}
\title{dec.curve}
\usage{
dec.curve(mat_input, mag, plot = TRUE)
}
\arguments{
\item{mat_input}{numeric: a kx3 matrix}

\item{mag}{numeric: how many times will be divided by the number of initial points}

\item{plot}{logical: if TRUE will be plotted the starting and final point matrices}
}
\value{
matt numeric: a kx3 matrix with points coordinates
}
\description{
This function computes the order of points on a open 3D curve and finds intermediate points
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
