% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Del_sp.R
\docType{data}
\name{Del_sp}
\alias{Del_sp}
\title{example dataset}
\usage{
data(Del_sp)
}
\description{
mesh coming from a ct-scan of a delphinidae
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
