% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Den_obs.R
\docType{data}
\name{Den_obs}
\alias{Den_obs}
\title{example dataset}
\usage{
data(Den_obs)
}
\description{
mesh coming from a ct-scan of Dendragapus obscurus
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
