% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_base_sur.R
\docType{data}
\name{DM_base_sur}
\alias{DM_base_sur}
\title{example dataset}
\usage{
data(DM_base_sur)
}
\description{
3D mesh of the first part of the Homo sapiens disarticulated model
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
