% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_face_sur.R
\docType{data}
\name{DM_face_sur}
\alias{DM_face_sur}
\title{example dataset}
\usage{
data(DM_face_sur)
}
\description{
3D mesh of the second part of the Homo sapiens disarticulated model
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
