% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_set.R
\docType{data}
\name{DM_set}
\alias{DM_set}
\title{example dataset}
\usage{
data(DM_set)
}
\description{
Landmark configurations of the two part of the disarticulated model
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
\keyword{Arothron}
