% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta.r
\name{dta}
\alias{dta}
\title{dta}
\usage{
dta(RM_sample, mod_1, mod_2, pairs_1, pairs_2, DM_mesh_1, DM_mesh_2, DM_set_1,
  DM_set_2, method = c("euclidean"))
}
\arguments{
\item{RM_sample}{3D array: 3D landmark configurations of the reference sample}

\item{mod_1}{numeric vector: vector containing the position of which landmarks belong to the first module}

\item{mod_2}{numeric vector: vector containing the position of which landmarks belong to the second module}

\item{pairs_1}{matrix: a X x 2 matrix containing the indices of right and left landmarks of the first module}

\item{pairs_2}{matrix: a X x 2 matrix containing the indices of right and left landmarks of the second module}

\item{DM_mesh_1}{mesh3d: mesh of the disarticulated model (first module)}

\item{DM_mesh_2}{mesh3d: mesh of the disarticulated model (second module)}

\item{DM_set_1}{matrix: 3D landmark set of the first module acquired on the disarticulated model}

\item{DM_set_2}{matrix: 3D landmark set of the second module acquired on the disarticulated model}

\item{method}{character: specify method to be used to individuate the best DTA ("euclidean" or "procrustes")}
}
\value{
AM_mesh mesh3d: mesh of the aligned model

AM_set matrix: landmark configuration of the aligned model

AM_id character: name of the item of the reference sample resulted as best DTA

AM_SF_1 numeric: scale factor used to scale the reference set (first module)

AM_SF_2 numeric: scale factor used to scale the reference set (second module)

distance numeric: distance between the landmark configuration of the aligned and the reference model

tot_proc numeric vector: procrustes distances between aligned and reference models (all DTAs)

tot_eucl numeric vector: euclidean distances between aligned and reference models (all DTAs)

setarray 3D array: landmark configurations of the disarticulated model aligned on each item of the reference sample
}
\description{
This function applyes the Digital Alignment Tool (DTA) on a disarticulated model using a reference sample
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
