% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endo_set.R
\docType{data}
\name{endo_set}
\alias{endo_set}
\title{example dataset}
\usage{
data(endo_set)
}
\description{
POVs defined inside the endocranial cavity
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
