% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.amira.path.r
\name{export.amira.path}
\alias{export.amira.path}
\title{export.amira.path}
\usage{
export.amira.path(vertices, filename, Lines = c(1:(dim(vertices)[1] - 1) - 1,
  -1), path = getwd())
}
\arguments{
\item{vertices}{numeric: a kx3 matrix}

\item{filename}{character: name of the requested output}

\item{Lines}{numeric: sequence of the vertices that defines the line}

\item{path}{character: folder path}
}
\description{
Convert and save a 3D matrix into a AmiraMesh ASCII Lineset (.am) object
}
\examples{
\dontrun{
x<-c(1:20)
y<-seq(1,3,length=20)
z<-rnorm(20,0.01)
vertices<-cbind(x,y,z)
example<-export.amira.path(vertices=vertices,filename="example_line") 
}
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
