% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_amira.r
\name{export_amira}
\alias{export_amira}
\title{export_amira}
\usage{
export_amira(lista, path = getwd())
}
\arguments{
\item{lista}{list containing 3D landmark sets}

\item{path}{character: path of the folder where saving the Amira landmark sets}
}
\description{
This function exports a list of 3D landmark set in separate files (format landmarkAscii)
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
