% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.mesh.rai.R
\name{ext.mesh.rai}
\alias{ext.mesh.rai}
\title{ext.mesh.rai}
\usage{
ext.mesh.rai(scans, mesh)
}
\arguments{
\item{scans}{an ext.int.mesh}

\item{mesh}{matrix mesh vertex (the same of the ext.int.mesh object)}
}
\description{
This function returns a 3D mesh with colours based on the vertices visibile from each point of view
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
