% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femsets.R
\docType{data}
\name{femsets}
\alias{femsets}
\title{example dataset}
\usage{
data(femsets)
}
\description{
3D semilandmark configurations of 21 human femora
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
\keyword{Arothron}
