% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_pov.R
\name{grid_pov}
\alias{grid_pov}
\title{grid_pov}
\usage{
grid_pov(mesh, expand = 1)
}
\arguments{
\item{mesh}{object of class mesh3d}

\item{expand}{numeric: scale factor for the grid for the interactive 3D plot}
}
\value{
matrice matrix: matrix with the x,y,z coordinates of the points of view
}
\description{
This function creates a grid for an interactive way to define the set of the points of view
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
