% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hom_sap.R
\docType{data}
\name{Hom_sap}
\alias{Hom_sap}
\title{example dataset}
\usage{
data(Hom_sap)
}
\description{
mesh coming from a ct-scan of Homo sapiens
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
