% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hum_ct.R
\docType{data}
\name{Hum_ct}
\alias{Hum_ct}
\title{example dataset}
\usage{
data(Hum_ct)
}
\description{
mesh coming from human ct scan
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
