% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hum_mr.R
\docType{data}
\name{Hum_mr}
\alias{Hum_mr}
\title{example dataset}
\usage{
data(Hum_mr)
}
\description{
mesh coming from human mri
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
