% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krd1_tooth.R
\docType{data}
\name{krd1_tooth}
\alias{krd1_tooth}
\title{example dataset}
\usage{
data(krd1_tooth)
}
\description{
3D mesh of a decidous Neanderthal tooth
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
