% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_frm2amira.R
\name{landmark_frm2amira}
\alias{landmark_frm2amira}
\title{landmark_frm2amira}
\usage{
landmark_frm2amira(path_folder_frm, path_amira_folder)
}
\arguments{
\item{path_folder_frm}{character: path of the folder where the .frm files are stored}

\item{path_amira_folder}{character: path folder to store the landmarkAscii configurations}
}
\description{
This function converts the .frm files, from Evan Toolbox, stored in a folder into the format landmarkAscii
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
