% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2array.R
\name{list2array}
\alias{list2array}
\title{list2array
converts a list into an array}
\usage{
list2array(mylist)
}
\arguments{
\item{mylist}{a list}
}
\value{
a kx3xn array with landmark coordinates
}
\description{
list2array
converts a list into an array
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Paolo Piras, Pasquale Raia
}
