% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listtoarray.R
\name{listtoarray}
\alias{listtoarray}
\title{listtoarray
convert a list into an array}
\usage{
listtoarray(mylist)
}
\arguments{
\item{mylist}{a list}
}
\value{
a kx3xn array with landmark coordinates
}
\description{
listtoarray
convert a list into an array
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
