% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lset2D_list.R
\docType{data}
\name{Lset2D_list}
\alias{Lset2D_list}
\title{example dataset}
\usage{
data(Lset2D_list)
}
\description{
List containing five 2D-landmark configurations acquired along five different anatomical views
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Paolo Piras, Pasquale Raia
}
\keyword{Arothron}
