% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lset3D_array.R
\docType{data}
\name{Lset3D_array}
\alias{Lset3D_array}
\title{example dataset}
\usage{
data(Lset3D_array)
}
\description{
Array containing a cranial 3D-landmark configuration acquired on a Primate sample
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Paolo Piras, Pasquale Raia
}
\keyword{Arothron}
