% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mac_fus.R
\docType{data}
\name{Mac_fus}
\alias{Mac_fus}
\title{example dataset}
\usage{
data(Mac_fus)
}
\description{
mesh coming from a ct-scan of Macaca fuscata
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
