% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malleus_bone.R
\docType{data}
\name{malleus_bone}
\alias{malleus_bone}
\title{example dataset}
\usage{
data(malleus_bone)
}
\description{
3D mesh of a human malleus
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
