% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAs_sets.R
\docType{data}
\name{MAs_sets}
\alias{MAs_sets}
\title{example dataset}
\usage{
data(MAs_sets)
}
\description{
Landmark configurations of the manual alignments
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
\keyword{Arothron}
