% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLECScorrection.R
\name{MLECScorrection}
\alias{MLECScorrection}
\title{MLECScorrection
Maximum Likelihood Estimation of the normalization factor to be applied to optimize the correlation between two landmark configurations to be combined by using twodviews and}
\usage{
MLECScorrection(array1, array2, scale = TRUE, nPCs = 1:5)
}
\arguments{
\item{array1}{array: first set of landmark configuration}

\item{array2}{array: second set of landmark configuration}

\item{scale}{logical: if FALSE the analysis is performed in the shape space, if TRUE the analysis is performed in the size and shape space (gpa without scaling)}

\item{nPCs}{numeric vector: specify which PC scores will be selected in the correlation test}
}
\value{
PCscores PCscores matrix of the combined dataset applying the normalization factor calculated by using the maximum likelihood estimation

PCs PCs matrix of the combined dataset applying the normalization factor calculated by using the maximum likelihood estimation

corr mean correlation between original and combined dataset

CSratios normalization factor calculated by using the maximum likelihood estimation
}
\description{
MLECScorrection
Maximum Likelihood Estimation of the normalization factor to be applied to optimize the correlation between two landmark configurations to be combined by using twodviews and
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
