% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCscoresCorr.R
\name{PCscoresCorr}
\alias{PCscoresCorr}
\title{PCscoresCorr
Perform a correlation test between two matrices of PCscores}
\usage{
PCscoresCorr(matrix1, matrix2, nPCs = 1:5)
}
\arguments{
\item{matrix1}{matrix: first set of PC scores}

\item{matrix2}{matrix: second set of PC scores}

\item{nPCs}{numeric vector: specify which PC scores will be selected in the correlation test}
}
\value{
corr the correlation values associated to each pair of PC scores

p.values p-values associated to the correlation test
}
\description{
PCscoresCorr
Perform a correlation test between two matrices of PCscores
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
