% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pov_selecter.R
\name{pov_selecter}
\alias{pov_selecter}
\title{pov_selecter}
\usage{
pov_selecter(mesh, grid, start.points = 250, method = "ast3d")
}
\arguments{
\item{mesh}{object of class mesh3d}

\item{grid}{matrix: a 3D grid}

\item{start.points}{numeric: number of center to be found}

\item{method}{character: select "a" or "c" for respectively AST-3D and CA-LSE method}
}
\value{
selection numeric: positioning vector of the selected points of the grid
}
\description{
Internal function to define the points of view
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
