% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri_sku_1.R
\docType{data}
\name{pri_sku_1}
\alias{pri_sku_1}
\title{example dataset}
\usage{
data(pri_sku_1)
}
\description{
list of 3D meshes containing 5 primates' skulls
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
