% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primate_skulls.R
\docType{data}
\name{primate_skulls}
\alias{primate_skulls}
\title{example dataset}
\usage{
data(primate_skulls)
}
\description{
list of 3D meshes containing 10 primates' skulls
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
