% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primendoR.R
\docType{data}
\name{primendoR}
\alias{primendoR}
\title{example dataset}
\usage{
data(primendoR)
}
\description{
right brain hemisphere of 19 primate species
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
\keyword{Arothron}
