% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rat_nor.R
\docType{data}
\name{Rat_nor}
\alias{Rat_nor}
\title{example dataset}
\usage{
data(Rat_nor)
}
\description{
mesh coming from a ct-scan of Rattus norvergicus
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
