% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.amira.dir.R
\name{read.amira.dir}
\alias{read.amira.dir}
\title{read.amira.dir}
\usage{
read.amira.dir(path.dir, nland)
}
\arguments{
\item{path.dir}{character: path of the folder}

\item{nland}{numeric: number of landmark sampled in Amira}
}
\value{
array.set numeric: a kx3xn array with landmark coordinates
}
\description{
This function reads and stores in an array the coordinated allocated in a folder in separate files (format landmarkAscii)
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
