% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.path.amira.R
\name{read.path.amira}
\alias{read.path.amira}
\title{read.path.amira}
\usage{
read.path.amira(path.name)
}
\arguments{
\item{path.name}{character: path of surface path .ascii extension file}
}
\value{
data numeric: a kxd matrix with xyz coordinates
}
\description{
This function extracts and orders the coordinate matrix from a surface path file from Amira
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
