% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{repmat}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{numeric: a matrix}

\item{m}{numeric: number of times to repeat the X matrix in row and column dimension}

\item{n}{numeric: repetition factor for each dimesion}
}
\value{
matrice: repeated matrix
}
\description{
This function repeats copies of a matrix
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
