% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMs_sets.R
\docType{data}
\name{RMs_sets}
\alias{RMs_sets}
\title{example dataset}
\usage{
data(RMs_sets)
}
\description{
Array containing the landmark coordinates of the reference sample for Digital Alignment Tool example
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
