% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP1.mesh.R
\docType{data}
\name{SCP1.mesh}
\alias{SCP1.mesh}
\title{example dataset}
\usage{
data(SCP1.mesh)
}
\description{
Mesh of the Saccopastore 1 Neanderthal skull
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
