% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinus_set.R
\docType{data}
\name{sinus_set}
\alias{sinus_set}
\title{example dataset}
\usage{
data(sinus_set)
}
\description{
POVs sampled inside the maxillary sinus cavity
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
