% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM_set.R
\docType{data}
\name{SM_set}
\alias{SM_set}
\title{example dataset}
\usage{
data(SM_set)
}
\description{
Landmark configuration associated to the starting model
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
\keyword{Arothron}
