% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical.flipping.R
\name{spherical.flipping}
\alias{spherical.flipping}
\title{spherical.flipping}
\usage{
spherical.flipping(C, mesh, param1, param2)
}
\arguments{
\item{C}{numeric: coordinates of the point of view}

\item{mesh}{object of class mesh3d}

\item{param1}{numeric: first parameter for spherical flipping (usually ranged from 0.1 to 3, try!)}

\item{param2}{numeric second paramter for spherical flipping (don't change it!)}
}
\description{
Internal spherical flippping function
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
