% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasf.mesh.R
\name{trasf.mesh}
\alias{trasf.mesh}
\title{trasf.mesh}
\usage{
trasf.mesh(mesh, barycenter)
}
\arguments{
\item{mesh}{a 3D mesh of class "mesh3d"}

\item{barycenter}{numeric: coordinates of the center}
}
\value{
mesh a 3D mesh of class "mesh3d"
}
\description{
This function centers a mesh on the barycenter coordinates
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
