% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twodviews.R
\name{twodviews}
\alias{twodviews}
\title{twodviews
Calculates the PCscores of the 2Dcomp matrix.  an array in a list storing each element of the third dimension of the array (specimen) as element of the list}
\usage{
twodviews(twodlist, scale = TRUE, vector = c(1:5))
}
\arguments{
\item{twodlist}{a list containing the landmark configurations of each anatomical view stored as separated lists}

\item{scale}{logical: TRUE for shape-space, FALSE for form-space}

\item{vector}{numeric vector: defines which views are to be used}
}
\value{
PCscores PC scores

PCs Pricipal Components (eigenvector matrix)

Variance table of the explained variance by the PCs

size vector containing the Centroid Size of each configuration

mshapes a list containing the mean shape of each 2D configuration

dims number of landmarks of each 2D configuration

twodlist the list used as input
}
\description{
twodviews
Calculates the PCscores of the 2Dcomp matrix.  an array in a list storing each element of the third dimension of the array (specimen) as element of the list
}
\examples{
library(Morpho)
#load the 2D primate dataset
data("Lset2D_list")
#combine the 2D datasets and PCA
combin2D<-twodviews(Lset2D_list,scale=TRUE,vector=c(1:5))
#plot of the first two Principal Components
plot(combin2D$PCscores)
text(combin2D$PCscores,labels=rownames(combin2D$PCscores))
#load the 3D primate dataset
data("Lset3D_array")
#GPA and PCA
GPA_3D<-procSym(Lset3D_array)
#plot of the first two Principal Components
plot(GPA_3D$PCscores)
text(GPA_3D$PCscores,labels=rownames(GPA_3D$PCscores))
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Paolo Piras, Pasquale Raia
}
