% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volendo.R
\name{volendo}
\alias{volendo}
\title{volendo}
\usage{
volendo(mesh, alpha_vol = 100, ncells = 1e+05)
}
\arguments{
\item{mesh}{object of class mesh3d}

\item{alpha_vol}{numeric: alpha shape for construction external concave hull}

\item{ncells}{numeric: approximative number of cell for 3D grid construction}
}
\value{
vol numeric: volume of the mesh expressed in cc
}
\description{
Calculate the volume of a mesh by using a voxel-based method
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Pasquale Raia
}
