% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoda_set.R
\docType{data}
\name{yoda_set}
\alias{yoda_set}
\title{example dataset}
\usage{
data(yoda_set)
}
\description{
Landmark set on Yoda
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
