% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoda_sur.R
\docType{data}
\name{yoda_sur}
\alias{yoda_sur}
\title{example dataset}
\usage{
data(yoda_sur)
}
\description{
Mesh of Yoda
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
\keyword{Arothron}
