% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arothron.R
\docType{package}
\name{Arothron-package}
\alias{Arothron-package}
\alias{Arothron}
\title{Geometric Morphometrics Analyses}
\description{
Tools for geometric morphometric analysis. The package includes tools of virtual anthropology to align two not articulated parts belonging to the same specimen, to build virtual cavities as endocast and to merge and combine the morphological information from 2D sets of coordinates acquired along different anatomical view. In addition we supply functions to import and export the coordinates of landmarks and 3D paths into 'landmarkAscii' and 'am' format files.
}
\author{
Antonio Profico, Alessio Veneziano, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia
}
