% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arothron.R
\docType{package}
\name{Arothron-package}
\alias{Arothron-package}
\alias{Arothron}
\title{eometric Morphometric Methods and Virtual Anthropology Tools}
\description{
Tools for geometric morphometric analysis. The package includes tools of virtual anthropology to align two not articulated parts belonging to the same specimen, to build virtual cavities as endocast (Profico et al, 2018 <doi:10.1002/ajpa.23493>).
}
\author{
Antonio Profico, Costantino Buzi, Silvia Castiglione, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
