% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_amira.R
\name{export_amira}
\alias{export_amira}
\title{export_amira}
\usage{
export_amira(lista, path)
}
\arguments{
\item{lista}{list containing 3D landmark sets}

\item{path}{character: path of the folder where saving the Amira landmark sets}
}
\description{
This function exports a list of 3D landmark set in separate files (format landmarkAscii)
}
\examples{
x<-c(1:20)
y<-seq(1,3,length=20)
z<-rnorm(20,0.01)
vertices<-cbind(x,y,z)
set<-list(vertices)
example<-export_amira(set,path=tempdir()) 
}
\author{
Antonio Profico, Alessio Veneziano, Marina Melchionna, Pasquale Raia
}
