% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_skull.R
\docType{data}
\name{human_skull}
\alias{human_skull}
\title{example dataset}
\usage{
data(human_skull)
}
\description{
3D mesh of a human skull
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
\keyword{Arothron}
