% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical.flipping.R
\name{spherical.flipping}
\alias{spherical.flipping}
\title{spherical.flipping}
\usage{
spherical.flipping(C, mesh, param1, param2)
}
\arguments{
\item{C}{numeric: coordinates of the point of view}

\item{mesh}{object of class mesh3d}

\item{param1}{numeric: first parameter for spherical flipping (usually ranged from 0.1 to 3, try!)}

\item{param2}{numeric second paramter for spherical flipping (don't change it!)}
}
\description{
Internal spherical flippping function
}
\references{
Profico, A., Schlager, S., Valoriani, V., Buzi, C., Melchionna, M., Veneziano, A., ... & Manzi, G. (2018). 
Reproducing the internal and external anatomy of fossil bones: Two new automatic digital tools. American Journal of Physical Anthropology, 166(4), 979-986.#' @export

Katz, S., Tal, A., & Basri, R. (2007). Direct visibility of point sets. In ACM SIGGRAPH 2007 papers (pp. 24-es).
}
\author{
Antonio Profico, Costantino Buzi, Marina Melchionna, Paolo Piras, Pasquale Raia, Alessio Veneziano
}
